Split up animations
